refactor: move errors to error.rs
