+++ crates/backends/src/error.rs
@@ -80,3 +80,49 @@ pub enum RestErrorKind {
     /// joining URL failed on: {0:?}
     JoiningUrlFailed(url::ParseError),
 }
+
+/// [`LocalBackendErrorKind`] describes the errors that can be returned by an action on the filesystem in Backends
+#[derive(Error, Debug, Display)]
+pub enum LocalBackendErrorKind {
+    /// directory creation failed: `{0:?}`
+    DirectoryCreationFailed(#[from] std::io::Error),
+    /// querying metadata failed: `{0:?}`
+    QueryingMetadataFailed(std::io::Error),
+    /// querying WalkDir metadata failed: `{0:?}`
+    QueryingWalkDirMetadataFailed(walkdir::Error),
+    /// executtion of command failed: `{0:?}`
+    CommandExecutionFailed(std::io::Error),
+    /// command was not successful for filename {file_name}, type {file_type}, id {id}: {status}
+    CommandNotSuccessful {
+        file_name: String,
+        file_type: String,
+        id: String,
+        status: ExitStatus,
+    },
+    /// error building automaton `{0:?}`
+    FromAhoCorasick(#[from] aho_corasick::BuildError),
+    /// {0:?}
+    FromSplitError(#[from] shell_words::ParseError),
+    /// {0:?}
+    #[error(transparent)]
+    FromTryIntError(#[from] TryFromIntError),
+    /// {0:?}
+    #[error(transparent)]
+    FromWalkdirError(#[from] walkdir::Error),
+    /// removing file failed: `{0:?}`
+    FileRemovalFailed(std::io::Error),
+    /// opening file failed: `{0:?}`
+    OpeningFileFailed(std::io::Error),
+    /// setting file length failed: `{0:?}`
+    SettingFileLengthFailed(std::io::Error),
+    /// can't jump to position in file: `{0:?}`
+    CouldNotSeekToPositionInFile(std::io::Error),
+    /// couldn't write to buffer: `{0:?}`
+    CouldNotWriteToBuffer(std::io::Error),
+    /// reading file contents failed: `{0:?}`
+    ReadingContentsOfFileFailed(std::io::Error),
+    /// reading exact length of file contents failed: `{0:?}`
+    ReadingExactLengthOfFileFailed(std::io::Error),
+    /// failed to sync OS Metadata to disk: `{0:?}`
+    SyncingOfOsMetadataFailed(std::io::Error),
+}
